+++ TPD_teleportDialog/GUI.hpp
@@ -41,15 +41,21 @@ class TPD_Teleport
       text = "$STR_STATE_MOVE";
       x = 0.5 - DIALOG_W / 2 * GRID_W + GRID_W;
       y = 66 * GRID_H;
-      w = DIALOG_W / 2 * GRID_W - 2 * GRID_W;
+      w = DIALOG_W / 3 * GRID_W - 2 * GRID_W;
       h = CTRL_DEFAULT_H;
       onButtonClick = "['teleport', nil, _this select 0] spawn TPD_fnc_teleport";
     };
+    class Preview: Teleport
+    {
+      text = "Preview Pos.";
+      x = 0.5 + DIALOG_W / 3 * GRID_W - DIALOG_W / 2 * GRID_W + GRID_W;
+      onButtonClick = "['previewPosition', nil, _this select 0] call TPD_fnc_teleport";
+    };
     class Close: Teleport
     {
-      idc = 1;
+      idc= 1;
       text = "$STR_DISP_OPT_CLOSE";
-      x = 0.5 + DIALOG_W / 2 * GRID_W - DIALOG_W / 2 * GRID_W + GRID_W;
+      x = 0.5 + DIALOG_W / 2 * GRID_W - DIALOG_W / 3 * GRID_W + GRID_W;
       onButtonClick = "";
     };
   };