+++ CHANGELOG.md
@@ -9,3 +9,4 @@
 - Added the ability to create custom locations
 - Easier setup of teleport actions via function
 - Teleportation to players which are in a vehicle is not possible anymore
+- Ability to preview position before teleporting